Split simulator support out of large files
