+++ Sources/CommandPalette/ContentView+CommandPaletteSurfaceMetadata.swift
@@ -1,0 +1,33 @@
+import Foundation
+
+extension ContentView {
+    func commandPaletteSurfaceKindLabel(for panelType: PanelType) -> String {
+        switch panelType {
+        case .terminal:
+            return String(localized: "commandPalette.kind.terminal", defaultValue: "Terminal")
+        case .browser:
+            return String(localized: "commandPalette.kind.browser", defaultValue: "Browser")
+        case .markdown:
+            return String(localized: "commandPalette.kind.markdown", defaultValue: "Markdown")
+        case .filePreview:
+            return String(localized: "commandPalette.kind.filePreview", defaultValue: "File Preview")
+        case .simulator:
+            return String(localized: "commandPalette.kind.simulator", defaultValue: "iOS Simulator")
+        }
+    }
+
+    func commandPaletteSurfaceKeywords(for panelType: PanelType) -> [String] {
+        switch panelType {
+        case .terminal:
+            return ["terminal", "shell", "console"]
+        case .browser:
+            return ["browser", "web", "page"]
+        case .markdown:
+            return ["markdown", "note", "preview"]
+        case .filePreview:
+            return ["file", "preview", "text", "pdf", "image"]
+        case .simulator:
+            return ["simulator", "ios", "iphone", "ipad"]
+        }
+    }
+}
